refactor: split from Extension.swift to Extensions+UIKit
Extension 파일 내부의 속성을 분리
